test: 학습 테스트 추가
