+++ src/test/java/com/h3/reservation/slackcalendar/domain/DateTimeTest.java
@@ -1,17 +1,14 @@
 package com.h3.reservation.slackcalendar.domain;
 
 import com.h3.reservation.slackcalendar.exception.InvalidTimeRangeException;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
-import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /**
  * @author heebg
@@ -59,29 +56,6 @@ class DateTimeTest {
         String endTime = "13:00";
 
         assertThrows(InvalidTimeRangeException.class, () -> DateTime.of(date, startTime, endTime));
-    }
-
-    @Test
-    @DisplayName("LocalTime 크기를 확인하는 학습테스트")
-    void local_time_later_than() {
-        LocalTime startTime = LocalTime.of(13, 0);
-        LocalTime endTime = LocalTime.of(13, 1);
-
-        assertTrue(startTime.isBefore(endTime));
-    }
-
-    @Test
-    @DisplayName("LocalDate get 포멧 확인하는 학습테스트")
-    void local_date_get_formatted() {
-        String date = "2019-12-10";
-        LocalDate localDate = LocalDate.parse(date);
-        String formattedDate = localDate.format(DateTimeFormatter.ofPattern("yyyy-MM-dd"));
-        assertEquals(formattedDate, date);
-
-        String time = "13:00";
-        LocalTime localTime = LocalTime.parse(time);
-        String formattedTime = localTime.format(DateTimeFormatter.ofPattern("HH:mm"));
-        assertEquals(formattedTime, time);
     }
 
     @Test
